+++ packages/login_client/test/login_client_test.dart
@@ -1,5 +1,6 @@
 import 'package:http/http.dart' show BaseRequest;
 import 'package:mockito/mockito.dart';
+import 'package:oauth2/oauth2.dart';
 import 'package:test/test.dart';
 
 import 'package:login_client/login_client.dart';